Fixes VSTS Bug 860954: Go To Definition causes a lengthy UI hang
https://devdiv.visualstudio.com/DevDiv/_workitems/edit/860954

Loading type system in background and removing the lock fixes the
issue. Locking is not needed because immutable data structures are
used now.
When opening a "blank" assembly browser the list of referenced
assemblies can be very long -esp. for .NET core projects it's better
not loading the referenced assemblies of all projects. It's a bit
unrelated but loadung such many assemblies can cause a very long load
time as well.
Even if it's not a hang it's basically "assembly browser does not do
anything" experience.